requirements.txt
aiohttp
aiohttp-socks
better-proxy
web3
PyExecJS
loguru